LABS/LAB-L-CICD-Orchestration/assets/DEVWKS-2176/requirements.txt
dnacentersdk==2.6.10
PyGithub==2.1.1
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.28.2
urllib3==1.26.14
